requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
feedparser==6.0.11
frozenlist==1.7.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
joblib==1.5.1
multidict==6.6.3
mysql-connector-python==9.4.0
numpy==2.3.2
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.34.2
sgmllib3k==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
threadpoolctl==3.6.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.1
